Dao jpa usuarios (#5)
* testeo de crud completo comida y menu

* inicio de Test y servicio con DAO para alumno

* alumnos DAO Test y avance en interfaz. faltan metodos

* dao alumno completado

* abstraccion a usuarioDAOHIbernateJPA para no repetir consultas sql

* abstraccion a genericos para usuarios

* ajuste en superclase de usuarioService para protected abstract de metodos que tienen que implemetnar las clases hijas

* usuarioDAO, admin, alumno y responsable completado con todos los testeos

* abstraccion a dao generico aplicado a productos comercializables y usuarios

+++ src/main/java/org/example/ttps2024grupo15/service/usuario/UsuarioService.java
@@ -1,12 +1,14 @@
 package org.example.ttps2024grupo15.service.usuario;
 
 
+import jakarta.persistence.NoResultException;
 import jakarta.transaction.Transactional;
 import org.example.ttps2024grupo15.dao.usuario.UsuarioDAO;
 import org.example.ttps2024grupo15.model.permiso.Rol;
 import org.example.ttps2024grupo15.model.request.usuario.UsuarioRequest;
 import org.example.ttps2024grupo15.model.usuario.Usuario;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public abstract class UsuarioService<T extends Usuario, S extends UsuarioDAO<T>, R extends UsuarioRequest> {
@@ -20,13 +22,26 @@ public abstract class UsuarioService<T extends Usuario, S extends UsuarioDAO<T>,
     @Transactional
     public T save(R usuarioRequest) {
         this.sanitizeRequest(usuarioRequest);
-        return this.dao.save(this.createUsuario(usuarioRequest));
+        try{
+            return this.dao.save(this.createUsuario(usuarioRequest));
+        } catch (Exception e){
+            e.printStackTrace();
+            throw new IllegalArgumentException("El usuario ya existe");
+        }
+
     }
 
     @Transactional
     public T update(Long id, R usuarioRequest){
+        this.validateID(id);
         this.sanitizeRequest(usuarioRequest);
-        T user = this.getUserById(id);
+        T user;
+        try{
+            user = this.dao.getById(id);
+        } catch (NoResultException e){
+            e.printStackTrace();
+            throw new IllegalArgumentException("El usuario no existe");
+        }
         user.setNombre(usuarioRequest.getNombre());
         user.setApellido(usuarioRequest.getApellido());
         user.setEmail(usuarioRequest.getEmail());
@@ -37,26 +52,74 @@ public abstract class UsuarioService<T extends Usuario, S extends UsuarioDAO<T>,
 
     @Transactional
     public void delete(Long id) {
-        this.dao.delete(id);
+        try{
+            this.validateID(id);
+            this.dao.delete(id);
+        } catch (NoResultException e){
+            e.printStackTrace();
+            throw new IllegalArgumentException("El usuario no existe");
+        }
     }
     @Transactional
     public void delete(T user) {
-        this.dao.delete(user);
+        try{
+            if(user == null){
+                throw new IllegalArgumentException("El usuario no puede ser nulo");
+            }
+            this.dao.delete(user);
+        } catch (NoResultException e){
+            e.printStackTrace();
+            throw new IllegalArgumentException("El usuario no existe");
+        }
     }
     public T getUserById(Long id) {
-        return dao.getById(id);
+        try{
+            this.validateID(id);
+            return dao.getById(id);
+        } catch (NoResultException e){
+            e.printStackTrace();
+            throw new IllegalArgumentException("El usuario no existe");
+        }
     }
 
     public List<T> getUsersByName(String name) {
-        return dao.getUsuarioPorNombre(name);
+        try{
+            this.validateStringInputParameter(name, "El nombre no puede ser nulo o vacio");
+            return dao.getUsuarioPorNombre(name);
+        } catch (NoResultException e){
+            e.printStackTrace();
+            return new ArrayList<>();
+        }
     }
 
     public List<T> getUsersByLastName(String lastName) {
-        return dao.getUsuarioPorApellido(lastName);
+        try{
+            this.validateStringInputParameter(lastName, "El apellido no puede ser nulo o vacio");
+            return dao.getUsuarioPorApellido(lastName);
+        } catch (NoResultException e){
+            e.printStackTrace();
+            return new ArrayList<>();
+        }
     }
 
     public T getUserByEmail(String email) {
-        return dao.getByEmail(email);
+        try{
+            this.validateStringInputParameter(email,  "El email no puede ser nulo o vacio");
+            return dao.getByEmail(email);
+        } catch (NoResultException e){
+            e.printStackTrace();
+            return null;
+        }
+    }
+
+    public T getUserByDni(Integer dni) {
+        try{
+            this.validateDNI(dni);
+            return dao.getByDni(dni);
+        } catch (NoResultException e){
+            e.printStackTrace();
+            return null;
+        }
     }
 
     public List<T> getAll() {
@@ -67,9 +130,7 @@ public abstract class UsuarioService<T extends Usuario, S extends UsuarioDAO<T>,
         return dao.getUsuariosPorRol(rol);
     }
 
-    protected abstract void sanitizeRequestSpecificFields(R usuarioRequest);
-
-    protected void sanitizeRequest(R usuarioRequest) {
+    private void sanitizeRequest(R usuarioRequest) {
         if(usuarioRequest.getNombre() == null || usuarioRequest.getNombre().isEmpty()){
             throw new IllegalArgumentException("El nombre no puede ser nulo o vacio");
         }
@@ -84,8 +145,25 @@ public abstract class UsuarioService<T extends Usuario, S extends UsuarioDAO<T>,
         }
         this.sanitizeRequestSpecificFields(usuarioRequest);
     }
-    protected abstract T createUsuario(R usuarioRequest);
-    protected void setUpdateSpecificFields(T user, R usuarioRequest) {
+    private void validateID(Long id) {
+        if(id == null){
+            throw new IllegalArgumentException("El id no puede ser nulo");
+        }
     }
+    private void validateDNI(Integer dni) {
+        if(dni == null){
+            throw new IllegalArgumentException("El dni no puede ser nulo");
+        }
+    }
+
+    private void validateStringInputParameter(String param, String message){
+        if(param == null || param.isEmpty()){
+            throw new IllegalArgumentException(message);
+        }
+    }
+    protected abstract T createUsuario(R usuarioRequest);
+    protected abstract void setUpdateSpecificFields(T user, R usuarioRequest) ;
+    protected abstract void sanitizeRequestSpecificFields(R usuarioRequest);
+
 }
